abstract out traverse configs
