+++ src/inline-loops.macro.js
@@ -1,5 +1,7 @@
 const { MacroError, createMacro } = require('babel-plugin-macros');
 const { createTemplates } = require('./templates');
+const { createTraverseConfigs } = require('./traverse');
+const { rename } = require('./utils');
 
 function myMacro({ references, babel }) {
   const { template, types: t } = babel;
@@ -33,71 +35,7 @@ function myMacro({ references, babel }) {
   } = references;
 
   const templates = createTemplates(template);
-
-  const stripReturnTraverseConfig = {
-    ReturnStatement(returnPath, returnState = {}) {
-      const arg = returnPath.get('argument');
-
-      if (returnState.isForEach) {
-        if (arg.node) {
-          const statement = arg.isExpression()
-            ? t.expressionStatement(arg.node)
-            : arg.node;
-
-          returnPath.insertBefore(statement);
-        }
-
-        returnPath.replaceWith(t.continueStatement());
-      } else {
-        returnPath.remove();
-      }
-    },
-  };
-
-  const bodyTraverseConfig = {
-    ArrayPattern(arrayPatternPath) {
-      arrayPatternPath.get('elements').forEach((element) => {
-        if (element.isIdentifier()) {
-          rename(element);
-        }
-      });
-    },
-    ObjectPattern(objectPatternPath) {
-      objectPatternPath.get('properties').forEach((property) => {
-        const value = property.get('value');
-
-        if (value.isIdentifier()) {
-          rename(value);
-        }
-      });
-    },
-    ReturnStatement(returnPath, returnState) {
-      const statementParent = returnPath.parentPath.getStatementParent();
-
-      if (
-        statementParent.isIfStatement() ||
-        statementParent.isSwitchStatement()
-      ) {
-        // If conditional returns exist, bail out inlining return statement.
-        returnState.value = null;
-        returnState.count = Infinity;
-        return;
-      }
-
-      returnState.value = returnState.count
-        ? null
-        : returnPath.get('argument').node;
-      returnState.count++;
-    },
-    ThisExpression(thisPath, thisState) {
-      thisState.containsThis = true;
-    },
-    VariableDeclarator(variablePath) {
-      const id = variablePath.get('id');
-
-      rename(id);
-    },
-  };
+  const traverseConfigs = createTraverseConfigs(babel);
 
   const handlers = {
     every: createHandleEverySome('every-left'),
@@ -206,7 +144,7 @@ function myMacro({ references, babel }) {
       value: null,
     };
 
-    body.traverse(bodyTraverseConfig, traverseState);
+    body.traverse(traverseConfigs.body, traverseState);
 
     const {
       containsThis: callbackContainsThis,
@@ -217,7 +155,7 @@ function myMacro({ references, babel }) {
     if (body.isBlockStatement()) {
       if (!callbackContainsThis) {
         if (returnCount < 2) {
-          body.traverse(stripReturnTraverseConfig, { isForEach });
+          body.traverse(traverseConfigs.stripReturn, { isForEach });
         }
 
         if (returnCount === 0) {
@@ -422,10 +360,6 @@ function myMacro({ references, babel }) {
     if (importedHandlerName) {
       handlers[importedHandlerName](path.get('callee'));
     }
-  }
-
-  function rename(path, newName) {
-    path.scope.rename(path.node.name, newName);
   }
 
   function replaceOrRemove(path, replacement) {
