+++ src/utils.js
@@ -1,0 +1,7 @@
+function rename(path, newName) {
+  path.scope.rename(path.node.name, newName);
+}
+
+module.exports = {
+  rename,
+};
